CompilerSpirV: Added support for user provided list of available extensions.
